[FRONT] Add push notif success,warning,danger to context
