+++ client/src/contexts/NotificationContext.js
@@ -1,6 +1,19 @@
 import React, {createContext, useState} from "react"
 import {Notification} from "../elements/Notification";
 
+const NOTIFICATION_LEVEL = {
+  DANGER: { duration: 8000, color: "#800000", bgColor: "#ff9290" },
+  WARNING: { duration: 4000, color: "#805500", bgColor: "#ffe8a3" },
+  SUCCESS: { duration: 5000, color: "#006500", bgColor: "#bcffb7" },
+};
+
+class NotificationModel {
+  constructor (message, level) {
+    this.message = message;
+    this.level = level;
+  }
+}
+
 const NotificationContext = createContext();
 
 const NotificationContextProvider = (props) => {
@@ -12,13 +25,30 @@ const NotificationContextProvider = (props) => {
     setState({notification: n, timer: setTimeout(removeNotification, n.level.duration)});
   };
 
+  const pushSuccess = (message) => {
+    const notification = new NotificationModel(message, NOTIFICATION_LEVEL.SUCCESS);
+    pushNotification(notification);
+  };
+
+  const pushWarning = (message) => {
+    const notification = new NotificationModel(message, NOTIFICATION_LEVEL.WARNING);
+    pushNotification(notification);
+  };
+
+  const pushDanger = (message) => {
+    const notification = new NotificationModel(message, NOTIFICATION_LEVEL.DANGER);
+    pushNotification(notification);
+  };
+
   const removeNotification = () => {
     setState({ notification: null, timer: null });
   };
 
   return (
     <NotificationContext.Provider value={{
-      pushNotification,
+      pushSuccess,
+      pushWarning,
+      pushDanger,
       removeNotification
     }}>
       { props.children }
